==> code/PREPROCESSING/5.web_of_science.R <==
# In the end this was canceled and we just did targeted searches from the browser.

devtools::install_github("Clarivate-SAR/woslite_r_client", force = T)

library(wosliterclient)

wosliterclient::


wosliterclient::ApiClient()
